feat(e2e): add visual tests for MDIconButton and FAB components
- implement visual regression tests for MDIconButton covering various states
- add visual tests for extended FAB interaction and state visuals
- include snapshot images for new tests to ensure visual consistency
