+++ tests/e2e/visual/shared-ui/md-button.spec.ts-snapshots/md-button-interaction-states-linux.png
