+++ tests/e2e/visual/shared-ui/md-button.spec.ts-snapshots/md-button-toggle-shapes-linux.png
