+++ tests/e2e/visual/shared-ui/md-fab-family.spec.ts-snapshots/md-extended-fab-interaction-states-linux.png
