+++ tests/e2e/visual/shared-ui/md-fab-family.spec.ts-snapshots/md-fab-size-comparison-linux.png
