+++ tests/e2e/visual/shared-ui/md-icon-button.spec.ts-snapshots/md-icon-button-toggle-interaction-states-linux.png
